requirements.txt
ipykernel
numpy
Pillow
pytesseract
pandas
jinja2
pdf2image
requests
beautifulsoup4
scikit-image
tqdm
matplotlib

datasets
transformers
sentencepiece

scikit-learn
statsmodels

# experiments
evaluate
torch
seqeval
accelerate
optuna
py-cpuinfo
igpu

#other
wordcloud
Unidecode